Changed to using a styled-component Library.
Everything  works the same as using inline or css file.
Benefit of using it is isolation or localisation of your CSS.
I am sure I do not like this since it results in mixing of your CSS and your code. I do not believe the benefit is worth potential troubles.